extract out isl_multi_no_domain_templ.c
These functions will also be needed for isl_multi_id,
implemented next.
Make them all static because they are only useful
for the template functions in isl_multi_*.
